pyproject.toml
[project]
name = "schwab-mcp"
version = "0.1.0"
description = "Schwab Model Context Protocol (MCP) for the Schwab API"
readme = "README.md"
authors = [
    { name = "Jason Kölker", email = "jason@koelker.net" }
]
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.11.0",
    "click>=8.3.0",

    "httpx>=0.27.2",
    "mcp>=1.17.0",
    "platformdirs>=4.5.0",
    "python-toon>=0.1.2",
    "pyyaml>=6.0.3",
    "schwab-py",
]

[project.scripts]
schwab-mcp = "schwab_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mcp[cli]>=1.17.0",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.0",
]
ta = [
    "pandas>=2.2.2",
    "pandas-ta-classic>=0.3.36",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=schwab_mcp --cov-report=term-missing"

[tool.coverage.run]
source = ["src/schwab_mcp"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
show_missing = true

[tool.pyright]
include = ["src", "tests"]
exclude = [
    "info",
    ".venv",
    "build",
    "dist",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
]
venvPath = "."
venv = ".venv"
reportMissingImports = false
typeCheckingMode = "basic"

[tool.uv.sources]
schwab-py = { git = "https://github.com/jkoelker/schwab-py.git", rev = "proxy_support" }
